refactor(agents): standardize management mutation payloads
