+++ src/app/api/agent-clients/[agentClientId]/credentials/route.ts
@@ -1,5 +1,9 @@
 import type { NextRequest } from "next/server";
-import { issueAgentCredentialForUser, issueAgentCredentialSchema } from "@/lib/agents";
+import {
+  issueAgentCredentialForUser,
+  issueAgentCredentialSchema,
+  listRecentAgentAuditEventsForUser,
+} from "@/lib/agents";
 import { AgentInputError, getErrorStatus, getPublicErrorMessage } from "@/lib/agents/errors";
 import { requireAuth, UnauthorizedError } from "@/lib/auth/require-auth";
 import { logger } from "@/lib/logger";
@@ -19,7 +23,8 @@ export async function POST(
     }
 
     const result = await issueAgentCredentialForUser(userId, agentClientId, parsed.data);
-    return Response.json(result, { status: 201 });
+    const recentEvents = await listRecentAgentAuditEventsForUser(userId);
+    return Response.json({ ...result, recentEvents }, { status: 201 });
   } catch (error) {
     if (error instanceof UnauthorizedError) {
       return Response.json({ error: "Unauthorized" }, { status: 401 });
@@ -33,4 +38,3 @@ export async function POST(
     return Response.json({ error: getPublicErrorMessage(error) }, { status });
   }
 }
-
