+++ src/app/api/agent-clients/[agentClientId]/revoke/route.ts
@@ -1,4 +1,4 @@
-import { revokeAgentClientForUser } from "@/lib/agents";
+import { listRecentAgentAuditEventsForUser, revokeAgentClientForUser } from "@/lib/agents";
 import { getErrorStatus, getPublicErrorMessage } from "@/lib/agents/errors";
 import { requireAuth, UnauthorizedError } from "@/lib/auth/require-auth";
 import { logger } from "@/lib/logger";
@@ -11,8 +11,9 @@ export async function POST(
     const { userId } = await requireAuth();
     const { agentClientId } = await context.params;
     const client = await revokeAgentClientForUser(userId, agentClientId);
+    const recentEvents = await listRecentAgentAuditEventsForUser(userId);
 
-    return Response.json({ client });
+    return Response.json({ client, recentEvents });
   } catch (error) {
     if (error instanceof UnauthorizedError) {
       return Response.json({ error: "Unauthorized" }, { status: 401 });
@@ -26,4 +27,3 @@ export async function POST(
     return Response.json({ error: getPublicErrorMessage(error) }, { status });
   }
 }
-
